% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{catalogues_available}
\alias{catalogues_available}
\title{Check for available catalogue datasets.}
\usage{
catalogues_available()
}
\value{
A data frame with information about available datasets, including dataset name, source,
type, genome, and file extension.
}
\description{
This function lists the available catalogue datasets included in the TCGAcatalogues package.
}
\examples{

# List available datasets
catalogues_available()

# Load datasets
catalogues_collection_acc <- catalogues_load('ACC')

# Load datasets as data.frames
catalogues_dataframe_acc <- catalogues_load('ACC', dataframe = TRUE)

}
