% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{decomp_available}
\alias{decomp_available}
\title{Check for available catalogue datasets.}
\usage{
decomp_available()
}
\value{
A data frame with information about available datasets, including dataset name, source,
type, genome, and file extension.
}
\description{
This function lists the available catalogue datasets included in the TCGAdecomp package.
}
\examples{

# List available datasets
decomp_available()

# Load datasets
decomp_collection_acc <- decomp_load('ACC')

# Load datasets as data.frames
decomp_dataframe_acc <- decomp_load('ACC', dataframe = TRUE)

}
