% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{decomp_load}
\alias{decomp_load}
\title{Load a decomposition dataset.}
\usage{
decomp_load(
  dataset,
  source = c("MC3", "Firehose"),
  type = c("SBS_96", "SBS_6", "SBS_1536", "DBS_78", "DBS_1248", "ID83"),
  genome = c("hg19", "hg38"),
  dataframe = FALSE
)
}
\arguments{
\item{dataset}{Name of the dataset to load.}

\item{source}{The source of the dataset (e.g., 'MC3' or 'Firehose').}

\item{type}{The type of decomposition (e.g., 'SBS_96' or 'SBS_6').}

\item{genome}{The genome version (e.g., 'hg19' or 'hg38').}

\item{dataframe}{Logical indicating whether to return the loaded dataset as a data frame (default: FALSE).}
}
\value{
A sigverse style decomposition collection (a list of data frames, each containing decomposition data for a unique sample).
If dataframe is TRUE, a single data frame with the entire dataset is returned.
}
\description{
This function loads a specific decomposition dataset based on the provided dataset name, source,
type, and genome. It can also return the loaded dataset as a data frame.
}
\examples{

# List available datasets
decomp_available()

# Load datasets
decomp_collection_acc <- decomp_load('ACC')

# Load datasets as data.frames
decomp_dataframe_acc <- decomp_load('ACC', dataframe = TRUE)

}
